Depth parameter dynamic

+++ wbn/errors.py
@@ -24,7 +24,7 @@ class InstanceCountError(WBNException):
 class MaxDepthExceededError(WBNException):
     """MaxDepthExceededError Exception."""
 
-    def __init__(self, depth: int):
+    def __init__(self, depth: float):
         self.depth = depth
 
     def __str__(self) -> str:
